Style calculator buttons and other elements
I did a lot of work spacing and styling the elements of the calculator
to make it look right. I changed the sizing and spacing of each
element to make it look like a calculator.

I did this in order to set the template for my javascript code. I will
add styling as needs arise, but this is a good starting place for what
comes next in development.

I did this by many css methods, but primarily using flex to space
out each element and making minor changes in padding and margin.