Create variables for display numbers
I created two variables for the display. The first one will hold the
numbers that the user inputs. The second is a variable that selects
the screen element and updates the text content to mirror the value
in the first variable.

I did this so I can keep track of the values that the user inputs
and immediately update the html.