Add delete button functionality
This commit adds functionality to the "delete" button on the
calculator. The getElementById method is used to assign the button to
the variable: deleteButton. The addEventListener method is used to
attach a click event to the deleteButton, which removes the last
character from the currentDisplayValue and updates the display with the
new value using textContent. This allows the user to easily modify the
current input and delete the last character entered.