Improve clear and addition button functionality
Code was added to the clearDisplay function. When the button is pressed, firstNumber, operator, and secondNumber variables all return to empty. This improves the clear function by emptying out the variables when it is pressed.

The plusSymbol event was improved as well. Using an if/else statement, the user can use the previous result of an equation as the first part of the next equation. The display will mirror this decision as well.

If the user also wants to put together a string of additions, the display will use the result of each calculation and update the display. the structure used for this button will be copied over to the other functions to complete the calculator's functionality.