Fix delete button bug after equation
The delete button only takes a string as an input. The bug appeared when the user tried to use the delete button after an equation. After an equation, the data type is a number, so this code turns the current display value to a string before performing the delete function.