Fix equals issue and infinity bug
There were two bugs fixed with this commit.

The first bug appeared when the user pressed "=" when an equation had already been performed. The if/then loop won't allow anything to happen if the oldDisplayValue already includes "=". This will stop the calculator equating things when there are no arguments.

The second bug appeared when the user divides something by 0. This was fixed by using a sassy alert box (per Odin instructions) and clearing the display if "Infinity" is ever found in a calculation.