Fix decimal bug
This commit fixes a bug that caused every number created by the operate() function to have two decimal places, even if it wasn't a decimal.

This commit causes operate() to check if the result is a decimal before rounding the result to the 2nd decimal place.

This will make math with the calculator more simple.