Fix various bugs related to equations with 0
There were several bugs related to the number 0. The first bug appeared when NaN came up as a result of a decimal divided by zero. The commit fixes the issue by checking the text content of the display - if it is NaN, an alert is fired and the display is cleared.

The second bug appeared to make a 0 look invisible during an equation if it were the first number. The code now makes the 0 visible.